+++ README.md
@@ -1,1 +1,7 @@
 # ServiceMocker
+
+Need to spin up some endpoint to test your input quickly this is it, a Spring boot app that you can have up and running in seconds that you can send input too;
+
+From eclipse select ServiveMockerApplication.java and run as Application you can then hit endpoints like 
+
+http://localhost:8000/helloWorld
