+++ ServiceMocker/src/main/java/com/thebelfastcodeforge/servicemocker/controllers/HelloGoodbyeController.java
@@ -5,26 +5,27 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class ChainController {
+public class HelloGoodbyeController {
+	
+	@RequestMapping(value = "/helloWorld", method = RequestMethod.GET)
+	public String helloWorld() {
+		return "Hello World";
+	}
+	
+	@RequestMapping(value = "/goodbyeWorld", method = RequestMethod.GET)
+	public String goodbyeWorld() {
+		return "Goodbye World";
+	}
+
 	
 	@RequestMapping(value = "/helloOne", method = RequestMethod.GET)
 	public String helloOne() {
 		return "Hello One";
 	}
 	
-	@RequestMapping(value = "/goodbyeOne", method = RequestMethod.GET)
-	public String goodbyeOne() {
-		return "Goodbye One";
-	}
-
 	@RequestMapping(value = "/helloTwo", method = RequestMethod.GET)
 	public String helloTwo() {
 		return "Hello Two";
-	}
-	
-	@RequestMapping(value = "/goodbyeTwo", method = RequestMethod.GET)
-	public String goodbyeTwo() {
-		return "Goodbye Two";
 	}
 	
 	@RequestMapping(value = "/helloThree", method = RequestMethod.GET)
@@ -32,19 +33,9 @@ public class ChainController {
 		return "Hello Three";
 	}
 	
-	@RequestMapping(value = "/goodbyeThree", method = RequestMethod.GET)
-	public String goodbyeThree() {
-		return "Goodbye One";
-	}
-	
 	@RequestMapping(value = "/helloFour", method = RequestMethod.GET)
 	public String helloFour() {
 		return "Hello Four";
-	}
-	
-	@RequestMapping(value = "/goodbyeFour", method = RequestMethod.GET)
-	public String goodbyeFour() {
-		return "Goodbye Four";
 	}
 	
 	@RequestMapping(value = "/helloFive", method = RequestMethod.GET)
@@ -52,9 +43,30 @@ public class ChainController {
 		return "Hello Five";
 	}
 	
+	@RequestMapping(value = "/goodbyeOne", method = RequestMethod.GET)
+	public String goodbyeOne() {
+		return "Goodbye One";
+	}
+	
+	@RequestMapping(value = "/goodbyeTwo", method = RequestMethod.GET)
+	public String goodbyeTwo() {
+		return "Goodbye Two";
+	}
+	
+	@RequestMapping(value = "/goodbyeThree", method = RequestMethod.GET)
+	public String goodbyeThree() {
+		return "Goodbye Three";
+	}
+	
+	@RequestMapping(value = "/goodbyeFour", method = RequestMethod.GET)
+	public String goodbyeFour() {
+		return "Goodbye Four";
+	}
+	
 	@RequestMapping(value = "/goodbyeFive", method = RequestMethod.GET)
 	public String goodbyeFive() {
 		return "Goodbye Five";
 	}
+
 
 }
