--- ServiceMocker/src/main/java/com/thebelfastcodeforge/servicemocker/controllers/PersonController.java
@@ -1,15 +1,0 @@
-package com.thebelfastcodeforge.servicemocker.controllers;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class PersonController {
-
-	
-
-	/*
-	 * @GetMapping("/getPerson")
-	 * 
-	 * @ResponseBody public String getPerson() { return "hello from getPerson"; }
-	 */
-}
